+++ runtime/string.c
@@ -1,10 +1,8 @@
-// String library implementation for Ripple C runtime
-// Uses for loops to avoid compiler issues with compound conditions in while loops
+#include <stddef.h>
+#include <string.h>
 
-// String copy - copies src to dst including null terminator
-// Returns dst
-char* strcpy(char* dst, char* src) {
-    int i;
+char *strcpy(char *dst, char *src) {
+    size_t i;
     for (i = 0; ; i++) {
         dst[i] = src[i];
         if (!src[i]) break;
@@ -12,14 +10,12 @@ char* strcpy(char* dst, char* src) {
     return dst;
 }
 
-// String copy with length limit
-char* strncpy(char* dst, char* src, int n) {
-    int i;
+char *strncpy(char *dst, char *src, size_t n) {
+    size_t i;
     for (i = 0; i < n; i++) {
         if (src[i]) {
             dst[i] = src[i];
         } else {
-            // Pad with nulls
             for (; i < n; i++) {
                 dst[i] = 0;
             }
@@ -29,18 +25,15 @@ char* strncpy(char* dst, char* src, int n) {
     return dst;
 }
 
-// String length
-int strlen(char* str) {
-    int len;
+size_t strlen(char *str) {
+    size_t len;
     for (len = 0; str[len]; len++) {
-        // Just counting
     }
     return len;
 }
 
-// String compare
-int strcmp(char* s1, char* s2) {
-    int i;
+int strcmp(char *s1, char *s2) {
+    size_t i;
     for (i = 0; ; i++) {
         if (!s1[i] || !s2[i]) break;
         if (s1[i] != s2[i]) {
@@ -50,9 +43,8 @@ int strcmp(char* s1, char* s2) {
     return s1[i] - s2[i];
 }
 
-// String compare with length limit
-int strncmp(char* s1, char* s2, int n) {
-    int i;
+int strncmp(char *s1, char *s2, size_t n) {
+    size_t i;
     for (i = 0; i < n; i++) {
         if (!s1[i] || !s2[i]) {
             return s1[i] - s2[i];
@@ -64,10 +56,9 @@ int strncmp(char* s1, char* s2, int n) {
     return 0;
 }
 
-// String concatenate
-char* strcat(char* dst, char* src) {
-    int dst_len = strlen(dst);
-    int i;
+char *strcat(char *dst, char *src) {
+    size_t dst_len = strlen(dst);
+    size_t i;
     for (i = 0; ; i++) {
         dst[dst_len + i] = src[i];
         if (!src[i]) break;
@@ -75,10 +66,9 @@ char* strcat(char* dst, char* src) {
     return dst;
 }
 
-// String concatenate with length limit
-char* strncat(char* dst, char* src, int n) {
-    int dst_len = strlen(dst);
-    int i;
+char *strncat(char *dst, char *src, size_t n) {
+    size_t dst_len = strlen(dst);
+    size_t i;
     for (i = 0; i < n; i++) {
         if (!src[i]) break;
         dst[dst_len + i] = src[i];
@@ -87,59 +77,53 @@ char* strncat(char* dst, char* src, int n) {
     return dst;
 }
 
-// Find character in string
-char* strchr(char* str, int c) {
-    int i;
+char *strchr(char *str, int c) {
+    size_t i;
     for (i = 0; str[i]; i++) {
         if (str[i] == c) {
-            return (char*)&str[i];
+            return (char *)&str[i];
         }
     }
     if (c == 0) {
-        return (char*)&str[i];  // Return pointer to null terminator
+        return (char *)&str[i];
     }
-    return 0;  // NULL
+    return 0;
 }
 
-// Find last occurrence of character in string
-char* strrchr(char* str, int c) {
-    char* last = 0;
-    int i;
+char *strrchr(char *str, int c) {
+    char *last = 0;
+    size_t i;
     for (i = 0; str[i]; i++) {
         if (str[i] == c) {
-            last = (char*)&str[i];
+            last = (char *)&str[i];
         }
     }
     if (c == 0) {
-        return (char*)&str[i];  // Return pointer to null terminator
+        return (char *)&str[i];
     }
     return last;
 }
 
-// Memory copy
-void* memcpy(void* dst, void* src, unsigned int n) {
-    char* d = (char*)dst;
-    char* s = (char*)src;
-    unsigned int i;
+void *memcpy(void *dst, void *src, size_t n) {
+    char *d = (char *)dst;
+    char *s = (char *)src;
+    size_t i;
     for (i = 0; i < n; i++) {
         d[i] = s[i];
     }
     return dst;
 }
 
-// Memory move (handles overlapping regions)
-void* memmove(void* dst, void* src, unsigned int n) {
-    char* d = (char*)dst;
-    char* s = (char*)src;
-    unsigned int i;
-    
+void *memmove(void *dst, void *src, size_t n) {
+    char *d = (char *)dst;
+    char *s = (char *)src;
+    size_t i;
+
     if (d < s || d >= s + n) {
-        // No overlap or dst before src - copy forward
         for (i = 0; i < n; i++) {
             d[i] = s[i];
         }
     } else {
-        // Overlap with dst after src - copy backward
         i = n;
         while (i) {
             i = i - 1;
@@ -149,21 +133,19 @@ void* memmove(void* dst, void* src, unsigned int n) {
     return dst;
 }
 
-// Memory set
-void* memset(void* dst, int c, unsigned int n) {
-    char* d = (char*)dst;
-    unsigned int i;
+void *memset(void *dst, int c, size_t n) {
+    char *d = (char *)dst;
+    size_t i;
     for (i = 0; i < n; i++) {
         d[i] = (char)c;
     }
     return dst;
 }
 
-// Memory compare
-int memcmp(void* s1, void* s2, unsigned int n) {
-    char* p1 = (char*)s1;
-    char* p2 = (char*)s2;
-    unsigned int i;
+int memcmp(void *s1, void *s2, size_t n) {
+    char *p1 = (char *)s1;
+    char *p2 = (char *)s2;
+    size_t i;
     for (i = 0; i < n; i++) {
         if (p1[i] != p2[i]) {
             return p1[i] - p2[i];
